+++ packages/adminui-vue/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
  - `UFile`: added preview mode for `html` (text/html) and `javascript`(application/def, application/javascript) documents 
  - `UCodeMirror`: added `src` property to preview script content in read-only mode
+ - `UCodeMirror` - `requestFullscreen()` method added - used by UFile to expand component to full screen
 
 ### Changed